feat(tools): implement Cypher queries for all MCP graph analysis tools
Functional description:
Enable LLMs to query the Neo4j code graph through functional MCP tools.
Previously, all tool handlers returned empty stub data. Now they execute
real Cypher queries to provide meaningful code analysis results.

Technical tasks completed:
- Implement search_nodes with regex pattern matching and type filtering
- Implement get_callers with variable-depth traversal for call graph analysis
- Implement get_callees with forward traversal from source functions
- Implement get_neighbors with bidirectional dependency discovery (USES/EXTENDS/IMPLEMENTS)
- Implement get_implementations for direct and indirect interface implementors
- Implement get_impact combining callers, dependents, implementors, and children analysis
- Implement find_path using Neo4j shortestPath algorithm
- Implement get_file_symbols with visibility filtering
- Update PLAN_DEVELOPMENT.md to mark MCP Tools step as completed

+++ mcp-server/src/tools/find-path/definition.ts
@@ -7,7 +7,7 @@ export const findPathDefinition = {
   inputSchema: {
     from_node: z.string().describe('Name of the starting node'),
     to_node: z.string().describe('Name of the target node'),
-    max_depth: z.number().min(1).max(10).optional().default(5).describe('Maximum path length to search'),
+    max_depth: z.number().int().min(1).max(10).optional().default(5).describe('Maximum path length to search'),
     relationship_types: z.array(z.string()).optional().describe('Filter by relationship types (e.g., ["CALLS", "USES"])'),
   },
 };
