+++ mcp-server/src/tools/get-callees/definition.ts
@@ -7,6 +7,6 @@ export const getCalleesDefinition = {
   inputSchema: {
     function_name: z.string().describe('Name of the function to find callees for'),
     class_name: z.string().optional().describe('Class containing the function (for disambiguation)'),
-    depth: z.number().min(1).max(5).optional().default(2).describe('Trace depth (1 = direct callees only)'),
+    depth: z.number().int().min(1).max(5).optional().default(2).describe('Trace depth (1 = direct callees only)'),
   },
 };
