+++ mcp-server/src/tools/get-callers/definition.ts
@@ -7,7 +7,7 @@ export const getCallersDefinition = {
   inputSchema: {
     function_name: z.string().describe('Name of the function to find callers for'),
     class_name: z.string().optional().describe('Class containing the function (for disambiguation)'),
-    depth: z.number().min(1).max(5).optional().default(2).describe('Trace depth (1 = direct callers only)'),
+    depth: z.number().int().min(1).max(5).optional().default(2).describe('Trace depth (1 = direct callers only)'),
   },
 };
