+++ mcp-server/src/tools/get-impact/definition.ts
@@ -7,7 +7,7 @@ export const getImpactDefinition = {
   inputSchema: {
     node_name: z.string().describe('Name of the node to analyze impact for'),
     node_type: z.enum(['class', 'interface', 'function', 'property']).optional().describe('Type of node (for disambiguation)'),
-    depth: z.number().min(1).max(10).optional().default(3).describe('Analysis depth'),
+    depth: z.number().int().min(1).max(10).optional().default(3).describe('Analysis depth'),
   },
 };
